+++ src/proto.h
@@ -237,6 +237,7 @@ void mbyte_im_set_active(int active_arg);
 #  include "vim9compile.pro"
 #  include "vim9execute.pro"
 #  include "vim9script.pro"
+#  include "vim9type.pro"
 # endif
 # include "window.pro"
